moves Leaderboard over to Tanstack server fns
